Fixing more problems that occurred with the 20110101.* releases.
*/makefile.in: automatic changes transferred from makefile.am and configure.ac.

AUTHORS: only date.
gpl_license.sh: run configure in subdirectory that gets removed.
build.sh: minor spelling error.
configure: changes automatically transferred from configure.ac.
configure.ac: automatic detection and handeling of ar missing.
config.h: only date.
configure.hpp: only date.
makefile.am: conditional makefile depending on CppAD_HAVE_AR.
whats_new_11.omh: Missing from previous commits.
makefile.am: use new CXX_FLAGS_FADBAD (defined by configure.ac).

git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1813 0e562018-b8fb-0310-a933-ecf073c7c197

+++ bin/gpl_license.sh
@@ -1,7 +1,7 @@
 #! /bin/bash -e
 # $Id$
 # -----------------------------------------------------------------------------
-# CppAD: C++ Algorithmic Differentiation: Copyright (C) 2003-10 Bradley M. Bell
+# CppAD: C++ Algorithmic Differentiation: Copyright (C) 2003-11 Bradley M. Bell
 #
 # CppAD is distributed under multiple licenses. This distribution is under
 # the terms of the
@@ -10,38 +10,38 @@
 # A copy of this license is included in the COPYING file of this distribution.
 # Please visit http://www.coin-or.org/CppAD/ for information on other licenses.
 # -----------------------------------------------------------------------------
+# Bradley M. Bell has given COIN-OR permission to use this script to generate 
+# a distribution of CppAD that has "GNU General Public License Version 2"
+# in place of "Common Public License Version 1.0." in all occurrences
+# of the message above.
+# -----------------------------------------------------------------------------
 if [ $0 != "bin/gpl_license.sh" ]
 then
 	echo "bin/gpl_license.sh: must be executed from its parent directory"
 	exit 1
 fi
-# -----------------------------------------------------------------------------
-#
-# Bradley M. Bell has given COIN-OR permission to use this script to generate 
-# a distribution of CppAD that has "GNU General Public License Version 2"
-# in place of "Common Public License Version 1.0." in all occurrences
-# of the message above.
 #
 version=`grep "^ *AC_INIT(" configure.ac | \
         sed -e 's/[^,]*, *\([^ ,]*\).*/\1/'`
 root_dir=`pwd`
 dir="cppad-$version"
 #
+# change into work sub-directory
 echo "cd work"
-cd work
+      cd work
 #
 # delete old version of *.gpl.tgz 
 if [ -e $dir.gpl.tgz ]
 then
 	echo "rm -f $dir.gpl.tgz"
-	rm -f $dir.gpl.tgz
+	      rm -f $dir.gpl.tgz
 fi
 #
 # delete old version of directory (if it exists)
 if [ -e $dir ]
 then
 	echo "rm -rf $dir"
-	rm -rf $dir
+	      rm -rf $dir
 fi
 if [ -e $dir ]
 then
@@ -51,11 +51,11 @@ fi
 #
 # extract from the *.cpl.tgz file
 echo "tar -xzf $dir.cpl.tgz"
-tar -xzf $dir.cpl.tgz
+      tar -xzf $dir.cpl.tgz
 #
-# make sure can change into new directory 
+# change into the work/cpl-distribution directory
 echo "cd $dir"
-cd $dir
+      cd $dir
 #
 # files that need changing
 list=`find . \
@@ -70,9 +70,9 @@ list=`find . \
 	\( -name '*.in'  \) -or \
 	\( -name '*.omh' \)`
 #
-# change back up to original directory (to be safe)
+# change back up to work directory (to be safe)
 echo "cd .."
-cd ..
+      cd ..
 #
 echo "Changing license from CPL to GPL"
 for file in $list
@@ -106,11 +106,11 @@ do
 	fi
 done
 #
-# change the COPYING file
+echo "change the COPYING file"
 sed -n -i $dir/COPYING -e '/-\{70\}/,/-\{70\}/p'
 cat $root_dir/gpl2.txt >> $dir/COPYING
 #
-# change the file cpl1.0.txt to the file gpl2.txt
+echo "change the file cpl1.0.txt to the file gpl2.txt"
 rm $dir/cpl1.0.txt
 cp $root_dir/gpl2.txt $dir/gpl2.txt
 #
@@ -135,25 +135,37 @@ do
 	fi
 done
 #
-# Rerun omhelp to change documentation version of license from CPL to GPL 
+# change into the work/gpl-distribution directory
 echo "cd $dir"
-cd $dir
+      cd $dir
 #
-echo "./configure > /dev/null"
-./configure > /dev/null
+# configure work/gpl-distribution/work so we can build the documentation
+# (This copies the output of configure to work/gpl-distribution.)
+if [ -d work ]
+then
+	echo "gpl_license.sh: There is a work subdirectory of the distribution"
+	exit 1
+fi
+echo "./build.sh configure"
+      ./build.sh configure 
 #
+echo "rm -r work"
+      rm -r work
+#
+# Now rebuild the documentation (so that it has GPL instead of CPL)
 echo "bin/run_omhelp.sh doc xml"
-bin/run_omhelp.sh doc xml
+      bin/run_omhelp.sh doc xml
 #
 if [ -e "doc/error.wrd" ]
 then
 	echo "rm -rf doc/error.wrd"
-	rm -rf doc/error.wrd
+	      rm -rf doc/error.wrd
 fi
 # 
+# change into the work directory
 echo "cd .."
-cd ..
+      cd ..
 #
-# create *.gpl.tgz file
+# create *.gpl.tgz file as copy or work/gpl-distribution directory
 echo "tar -czf $dir.gpl.tgz $dir"
-tar -czf $dir.gpl.tgz $dir
+      tar -czf $dir.gpl.tgz $dir
